golangci-lint: initialize
Initialize golangci-lint in the current setup.

Many of the 'linters' are not working due to the fact that we need to
set them up by calling preRun func.  However some of the more basic
linters just work out of the box.

We also use `go work sync` to manage a consistent set of dependencies
between staticcheck and golangci-lint. This allows the current workspace
to work with both, but it's not a recommended setting for others repos.
Folks who use other repos should try using a single go.mod file if
possible.