goci-lint: init a rewrite
goci-lint aimed to be a simpler rewrite of golangci-lint's collection of
analyzers.

The main goal is to remove how golangci-lint propagate configuration and
collect results to reduce bloat when loading an analyzer. This should
significantly reduce the I/O requirements of loading an analyzer and
speed up execution speed by a lot.

The secondary goal is to organize the packages in a way that user would
only need to consume the dependencies that they would need for that
particular analyzer and not all of `golangci-lint` dependencies.