gitignore: add go.work.sum
go.work.sum is created when a go command is needed to be run in a sub
module of the repository but the go.mod file of that sub directory was
incomplete.  In such case, go.work.sum file would be created to
temporarily hole the information needed by gopls.

Let's ignore this file for now as we would be expecting all
sub-modules to run 'go mod tidy' regularly.