gazelle: remove exclusion on tests dir
Now that we have a gazelle extension that helps generate filegroup
targets, we want to leverage it to generate targets for 'tests'
directories.

We accomplish this by removing the top-level BUILD's gazelle:exclude
directive.  Instead, we specify in each dir regarding which gazelle
language would we want to enable/disable using gazelle:lang directive.

Worth to note that this directive's value is propagate down to sub-dir,
so we also need to re-enable Go support for sub-dir if we have disabled
Go support in parent dir.

This also let us remove the rest of the gazelle:ignore directives that
we have been using.